build fix

+++ src/app/api/deleteProperty/[id]/route.ts
@@ -1,11 +1,22 @@
 import { NextResponse } from 'next/server';
 import Property from '../../addUserToProperty/Property';
+import { authenticate } from '@/lib/authenticate';
 export async function DELETE(
   request: Request,
   { params }: { params: { id: string } }
 ) {
   const propertyId = params.id;
-
+  const admin = await authenticate(request);
+  if (!admin) {
+    return Response.json(
+      {
+        message: 'No autorizado'
+      },
+      {
+        status: 401
+      }
+    );
+  }
   try {
     const deleted = await Property.findOneAndDelete({ _id: propertyId });
     return NextResponse.json(
